requirements.txt
asgiref==3.5.2
Django==3.1.7
django-rest-framework==0.1.0
django-simple-history==2.11.0
djangorestframework==3.13.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
